Adds MIT License to DontMissANote
This commit adds the MIT license to the DontMissANote project. From some
conversations w/ Dipak and Arun, it sounds like the intent
of this project was to be placed into the public for learning and use.
The MIT license fits well with that goal, mitigating liability and also
helping other people learn how to work w/ our APIs.